:bug: Do not update skipped tests on reload
In the tree, the state of a skipped test from a test run should remain the
same on reload.
The 'skipped' state has two meanings; A skipped test from a reload and
a skipped test from a test run. If a test is "unskipped from reload", it should
get the retired version of its state before skipping.